Added views.py
